Remove `polkadot-node-subsystem-util` crate, pull necessary code into `polkadot-node-collation-generation` crate
