+++ polkadot/node/collation-generation/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2021"
 futures = "0.3.21"
 tracing = "0.1.32"
 polkadot-node-subsystem = { path = "../subsystem" }
-polkadot-node-subsystem-util = { path = "../subsystem-util" }
 sp-core = { git = "https://github.com/paritytech/substrate", rev = "c364008a6c7da8456e17967f55edf51e45146998" }
 sp-maybe-compressed-blob  = { git = "https://github.com/paritytech/substrate", rev = "c364008a6c7da8456e17967f55edf51e45146998" }
 sp-runtime = { git = "https://github.com/paritytech/substrate", rev = "c364008a6c7da8456e17967f55edf51e45146998" }
@@ -17,4 +16,5 @@ parity-scale-codec = { version = "3.1.2", default-features = false, features = [
 
 cirrus-node-primitives = { path = "../../../crates/cirrus-node-primitives" }
 sp-executor  = { path = "../../../crates/sp-executor" }
+subspace-core-primitives = { path = "../../../crates/subspace-core-primitives" }
 subspace-runtime-primitives = { path = "../../../crates/subspace-runtime-primitives" }
