--- polkadot/node/subsystem-util/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "polkadot-node-subsystem-util"
-version = "0.9.13"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2021"
-description = "Subsystem traits and message definitions"
-
-[dependencies]
-async-trait = "0.1.52"
-futures = "0.3.21"
-pin-project = "1.0.10"
-thiserror = "1.0.30"
-tracing = "0.1.32"
-
-polkadot-node-subsystem = { package = "polkadot-node-subsystem", path = "../subsystem" }
-polkadot-overseer = { path = "../overseer" }
-
-sp-core = { git = "https://github.com/paritytech/substrate", rev = "c364008a6c7da8456e17967f55edf51e45146998" }
-sp-runtime = { git = "https://github.com/paritytech/substrate", rev = "c364008a6c7da8456e17967f55edf51e45146998" }
-
-sp-executor = { path = "../../../crates/sp-executor" }
-subspace-core-primitives = { path = "../../../crates/subspace-core-primitives" }
-subspace-runtime-primitives = { path = "../../../crates/subspace-runtime-primitives" }
